secure_step_backend/requirements.txt
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
tensorflow==2.15.0
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
librosa==0.10.1
scipy